dependency: Ensure the client can be used with the latest version of `typeguard` (#334)
This upgrade includes a breaking change in the `typeguard` API. `typeguard.check_type` doesn't rely on `typeguard.config` anymore. The solutions are either:

- Explicitely mention the config in the `typeguard.check_type`
- Rely on the `@typechecked` decorator

I went for the first one as we don't have many call to `typeguard.check_type`.

Note that version 4.0.0 was released on 2023-05-12 so there's no reason to try and maintain compatibility with versions < 4.